Allow configuration for `"<environment>": { ... lambda_name": "<name>" }`
This allows a configuration key `lambda_name` to specify a name for your AWS Lambda function. This is useful for environments where you only use zappa for CLI and serving, but not for deployment. For example: we use CodeBuild to deploy our code. We don't name our functions `<name>_<env>` the way zappa does. Zappa shouldn't be too opinionated on this. Let the user define, else fallback to your own presumed name format.